cleanup and loss function
